requirements.txt
streamlit
pandas
python-dotenv
crewai
crewai[tools]
langchain-community
gTTS
Flask==2.0.1
requests==2.26.0
oauthlib==3.1.1
pyOpenSSL==20.0.1